# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AniDB API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/codecs": {
            "get": {
                "tags": [
                    "Codec"
                ],
                "summary": "Get All Codecs",
                "operationId": "e0c231e244a324da7fa9837fe561cc2c",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "audio": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/CodecAudio"
                                                    }
                                                },
                                                "video": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/CodecVideo"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/codecs/audio": {
            "get": {
                "tags": [
                    "Codec"
                ],
                "summary": "Get All Audio Codecs",
                "operationId": "0a525c027792c43d27bfb8afa6a455fd",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CodecAudio"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Codec"
                ],
                "summary": "Add an Audio Codec",
                "operationId": "abf83d64014d56f89132113c8c1c0417",
                "parameters": [
                    {
                        "name": "codec",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Sample Codec"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/codecs/audio/{audio_codec_id}": {
            "put": {
                "tags": [
                    "Codec"
                ],
                "summary": "Edit an Audio Codec",
                "operationId": "62b34b01fcc30f147780312aa23c4504",
                "parameters": [
                    {
                        "name": "audio_codec_id",
                        "in": "path",
                        "description": "Audio Codec ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "codec",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Sample Codec"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Codec"
                ],
                "summary": "Delete an Audio Codec",
                "operationId": "2bdf95a30a93dcf0e0e7eea8b6cef18f",
                "parameters": [
                    {
                        "name": "audio_codec_id",
                        "in": "path",
                        "description": "Audio Codec ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/codecs/video": {
            "get": {
                "tags": [
                    "Codec"
                ],
                "summary": "Get All Video Codecs",
                "operationId": "4dd26c319592446b65eba6cc3d097957",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CodecVideo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Codec"
                ],
                "summary": "Add a Video Codec",
                "operationId": "8c456ec5babdf24cee4a7128fe096ab8",
                "parameters": [
                    {
                        "name": "codec",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Sample Codec"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/codecs/video/{video_codec_id}": {
            "put": {
                "tags": [
                    "Codec"
                ],
                "summary": "Edit a Video Codec",
                "operationId": "197b7187f4970d40e0f70ef9416087f6",
                "parameters": [
                    {
                        "name": "video_codec_id",
                        "in": "path",
                        "description": "Video Codec ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "codec",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Sample Codec"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Codec"
                ],
                "summary": "Delete an Video Codec",
                "operationId": "a1c38d740d3722e13edc21d1d4c802d3",
                "parameters": [
                    {
                        "name": "video_codec_id",
                        "in": "path",
                        "description": "Video Codec ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CodecAudio": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "codec": {
                        "type": "string"
                    },
                    "order": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "codec": "AAC 2.0",
                    "order": null
                }
            },
            "CodecVideo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "codec": {
                        "type": "string"
                    },
                    "order": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "codec": "x264 8bit",
                    "order": null
                }
            },
...
        },
        "responses": {
            "Success": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "format": "int32"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "status": 200,
                            "message": "Success"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "format": "int32"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "status": 401,
                            "message": "Unauthorized"
                        }
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}